ref: accelerator connector methods x/n (#3470)
